cobra/requirements/dev.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile --output-file=cobra/requirements/dev.txt cobra/requirements/dev.in
#
amqp==5.0.9
    # via
    #   -r cobra/requirements/base.txt
    #   kombu
asgiref==3.4.1
    # via
    #   -r cobra/requirements/base.txt
    #   django
    #   djoser
billiard==3.6.4.0
    # via
    #   -r cobra/requirements/base.txt
    #   celery
black==21.12b0
    # via -r cobra/requirements/dev.in
celery==5.2.1
    # via -r cobra/requirements/base.txt
certifi==2021.10.8
    # via
    #   -r cobra/requirements/base.txt
    #   requests
cffi==1.15.0
    # via
    #   -r cobra/requirements/base.txt
    #   cryptography
charset-normalizer==2.0.9
    # via
    #   -r cobra/requirements/base.txt
    #   requests
click==8.0.3
    # via
    #   -r cobra/requirements/base.txt
    #   black
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.0
    # via
    #   -r cobra/requirements/base.txt
    #   celery
click-plugins==1.1.1
    # via
    #   -r cobra/requirements/base.txt
    #   celery
click-repl==0.2.0
    # via
    #   -r cobra/requirements/base.txt
    #   celery
coreapi==2.3.3
    # via
    #   -r cobra/requirements/base.txt
    #   djangorestframework-stubs
    #   djoser
    #   drf-yasg
coreschema==0.0.4
    # via
    #   -r cobra/requirements/base.txt
    #   coreapi
    #   drf-yasg
coverage==6.2
    # via -r cobra/requirements/dev.in
cryptography==36.0.1
    # via
    #   -r cobra/requirements/base.txt
    #   social-auth-core
defusedxml==0.7.1
    # via
    #   -r cobra/requirements/base.txt
    #   python3-openid
    #   social-auth-core
django==4.0
    # via
    #   -r cobra/requirements/base.txt
    #   django-cors-headers
    #   django-extensions
    #   django-filter
    #   django-jazzmin
    #   django-stubs
    #   django-stubs-ext
    #   djangorestframework
    #   djangorestframework-simplejwt
    #   drf-yasg
django-cors-headers==3.10.1
    # via -r cobra/requirements/base.txt
django-extensions==3.1.5
    # via -r cobra/requirements/dev.in
django-filter==21.1
    # via -r cobra/requirements/base.txt
django-jazzmin==2.4.8
    # via -r cobra/requirements/base.txt
django-stubs==1.9.0
    # via
    #   -r cobra/requirements/dev.in
    #   djangorestframework-stubs
django-stubs-ext==0.3.1
    # via django-stubs
django-templated-mail==1.1.1
    # via
    #   -r cobra/requirements/base.txt
    #   djoser
djangorestframework==3.13.1
    # via
    #   -r cobra/requirements/base.txt
    #   djangorestframework-simplejwt
    #   drf-yasg
djangorestframework-simplejwt==4.8.0
    # via
    #   -r cobra/requirements/base.txt
    #   djoser
djangorestframework-stubs==1.4.0
    # via -r cobra/requirements/dev.in
djoser==2.1.0
    # via -r cobra/requirements/base.txt
drf-flex-fields==0.9.6
    # via -r cobra/requirements/base.txt
drf-yasg==1.20.0
    # via -r cobra/requirements/base.txt
factory-boy==3.2.1
    # via -r cobra/requirements/dev.in
faker==10.0.0
    # via
    #   -r cobra/requirements/dev.in
    #   factory-boy
flake8==4.0.1
    # via -r cobra/requirements/dev.in
freezegun==1.1.0
    # via -r cobra/requirements/dev.in
idna==3.3
    # via
    #   -r cobra/requirements/base.txt
    #   requests
inflection==0.5.1
    # via
    #   -r cobra/requirements/base.txt
    #   drf-yasg
isort==5.10.1
    # via -r cobra/requirements/dev.in
itypes==1.2.0
    # via
    #   -r cobra/requirements/base.txt
    #   coreapi
jinja2==3.0.3
    # via
    #   -r cobra/requirements/base.txt
    #   coreschema
kombu==5.2.2
    # via
    #   -r cobra/requirements/base.txt
    #   celery
markupsafe==2.0.1
    # via
    #   -r cobra/requirements/base.txt
    #   jinja2
mccabe==0.6.1
    # via flake8
mypy==0.930
    # via
    #   -r cobra/requirements/dev.in
    #   django-stubs
    #   djangorestframework-stubs
mypy-extensions==0.4.3
    # via
    #   black
    #   mypy
oauthlib==3.1.1
    # via
    #   -r cobra/requirements/base.txt
    #   requests-oauthlib
    #   social-auth-core
packaging==21.3
    # via
    #   -r cobra/requirements/base.txt
    #   drf-yasg
parameterized==0.8.1
    # via -r cobra/requirements/dev.in
pathspec==0.9.0
    # via black
platformdirs==2.4.0
    # via black
prompt-toolkit==3.0.24
    # via
    #   -r cobra/requirements/base.txt
    #   click-repl
psycopg2==2.9.2
    # via -r cobra/requirements/base.txt
pycodestyle==2.8.0
    # via flake8
pycparser==2.21
    # via
    #   -r cobra/requirements/base.txt
    #   cffi
pyflakes==2.4.0
    # via flake8
pyjwt==2.3.0
    # via
    #   -r cobra/requirements/base.txt
    #   djangorestframework-simplejwt
    #   social-auth-core
pyparsing==3.0.6
    # via
    #   -r cobra/requirements/base.txt
    #   packaging
python-dateutil==2.8.2
    # via
    #   faker
    #   freezegun
python3-openid==3.2.0
    # via
    #   -r cobra/requirements/base.txt
    #   social-auth-core
pytz==2021.3
    # via
    #   -r cobra/requirements/base.txt
    #   celery
    #   djangorestframework
requests==2.26.0
    # via
    #   -r cobra/requirements/base.txt
    #   coreapi
    #   djangorestframework-stubs
    #   requests-oauthlib
    #   social-auth-core
requests-oauthlib==1.3.0
    # via
    #   -r cobra/requirements/base.txt
    #   social-auth-core
ruamel.yaml==0.17.17
    # via
    #   -r cobra/requirements/base.txt
    #   drf-yasg
ruamel.yaml.clib==0.2.6
    # via
    #   -r cobra/requirements/base.txt
    #   ruamel.yaml
six==1.16.0
    # via
    #   -r cobra/requirements/base.txt
    #   click-repl
    #   python-dateutil
    #   social-auth-app-django
social-auth-app-django==4.0.0
    # via
    #   -r cobra/requirements/base.txt
    #   djoser
social-auth-core==4.1.0
    # via
    #   -r cobra/requirements/base.txt
    #   social-auth-app-django
sqlparse==0.4.2
    # via
    #   -r cobra/requirements/base.txt
    #   django
text-unidecode==1.3
    # via faker
toml==0.10.2
    # via django-stubs
tomli==1.2.3
    # via
    #   black
    #   mypy
types-freezegun==1.1.3
    # via -r cobra/requirements/dev.in
types-pytz==2021.3.3
    # via django-stubs
types-pyyaml==6.0.1
    # via django-stubs
typing-extensions==4.0.1
    # via
    #   black
    #   django-stubs
    #   django-stubs-ext
    #   djangorestframework-stubs
    #   mypy
uritemplate==4.1.1
    # via
    #   -r cobra/requirements/base.txt
    #   coreapi
    #   drf-yasg
urllib3==1.26.7
    # via
    #   -r cobra/requirements/base.txt
    #   requests
vine==5.0.0
    # via
    #   -r cobra/requirements/base.txt
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.5
    # via
    #   -r cobra/requirements/base.txt
    #   prompt-toolkit

# The following packages are considered to be unsafe in a requirements file:
# setuptools
